requirements.txt
flask
werkzeug
groq
pdfplumber
python-docx
python-dotenv